chore(api,lib): move referent schema to lib
