+++ api/src/models/index.ts
@@ -15,7 +15,7 @@ export { MissionModel, MissionDocument, MissionType } from "./mission";
 export { MissionAPIModel, MissionAPIDocument, MissionAPIType } from "./missionAPI";
 export { MissionEquivalenceModel, MissionEquivalenceDocument, MissionEquivalenceType } from "./missionEquivalence";
 export { ProgramModel, ProgramDocument, ProgramType } from "./program";
-export { ReferentModel, ReferentDocument, ReferentType } from "./referent";
+export { ReferentModel, ReferentDocument } from "./referent";
 export { SchoolModel, SchoolDocument, SchoolType } from "./school";
 export { SchoolRAMSESModel, SchoolRAMSESDocument, SchoolRAMSESType } from "./schoolRAMSES";
 export { SessionPhase1Model, SessionPhase1Document, SessionPhase1Type } from "./sessionPhase1";
