+++ packages/lib/src/mongoSchema/index.ts
@@ -4,3 +4,4 @@ export * from "./cle/classe";
 export * from "./cle/etablissement";
 export * from "./cohesionCenter";
 export * from "./pointDeRassemblement";
+export * from "./referent";
